feat: Add methods to handle GNSS and Network time on iOS
This commit adds two new methods, `getGnssTime` and `getNetworkTime`, to the `CapacitorRealTimePlugin` class in the iOS plugin file. These methods are used to handle requests for GNSS time and network time respectively. If called on iOS, they will reject the request with an appropriate error message indicating that these features are not available on iOS.