feat: Add network time synchronization functionality
This commit adds the necessary code changes to synchronize the app's time with a network time server. The `CapacitorRealTimePlugin` now includes a new method called `getTrueTime()` that retrieves the current network time using the Apache Commons Net library. The plugin also imports additional classes and libraries required for this functionality.

In the Android implementation, a new dependency on `commons-net:commons-net:3.10.0` is added in the build.gradle file.

In the iOS implementation, the Kronos framework is added as a pod dependency in both the Podfile and project.pbxproj files.

The web implementation of `CapacitorRealTimeWeb` includes a mock implementation for `getTrueTime()` that logs a message indicating it is not available in web environments.

These changes enable developers to retrieve accurate network time within their apps.